+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-dev:
+1.33.1:
   - add "slot" to "proposer duties" command
   - add activation epoch and time to "validator info" command where applicable
   - add "holesky" to the list of supported networks
+  - avoid crash when requesting validators from beacon node without debug enabled
 
 1.33.0:
   - show all slots with 'synccommittee inclusion'